chat_messenger/requirements.txt
ace_tools==0.0
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
apparmor==3.0.4
appdirs==1.4.4
apturl==0.5.2
asgiref==3.9.2
async-timeout==5.0.1
attrs==24.3.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
black==21.7b0
blinker==1.9.0
blis==0.7.11
Brlapi==0.8.3
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.16.0
channels==4.3.1
chardet==4.0.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudinary==1.36.0
cloudpathlib==0.16.0
colorama==0.4.4
command-not-found==0.3
confection==0.1.4
configobj==5.0.9
configparser==7.2.0
contourpy==1.2.1
crispy-bootstrap4==2022.1
cryptography==42.0.7
cssselect==1.3.0
cupshelpers==1.0
curl_cffi==0.13.0
cycler==0.12.1
cymem==2.0.8
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==0.5.0
Django==5.2.6
django-admin-interface==0.28.7
django-cloudinary-storage==0.3.0
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-crispy-forms==2.0
django-debug-toolbar==4.4.2
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.2
django-heroku==0.3.1
django-import-export==4.1.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
docker==7.1.0
docx2txt==0.8
drf-yasg==1.21.7
duplicity==0.8.21
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
environ==1.0
et_xmlfile==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fake-useragent==2.2.0
fasteners==0.14.1
filelock==3.16.1
fitz==0.0.1.dev2
fonttools==4.53.1
fpdf==1.7.2
frontend==0.0.3
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.5.0
future==0.18.2
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.44
googlesearch-python==1.2.5
greenlet==3.2.3
gunicorn==20.1.0
gyp==0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
ijson==3.4.0
importlib-metadata==4.6.4
importlib_resources==6.5.2
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
jiter==0.11.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
kiwisolver==1.4.5
langcodes==3.4.0
language-selector==0.1
language_data==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
LibAppArmor==3.0.4
lockfile==0.12.2
looseversion==1.3.0
louis==3.20.0
lxml==5.3.1
lxml_html_clean==0.4.2
macaroonbakery==1.3.1
Mako==1.1.3
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.1
mdurl==0.1.2
mistralai==1.6.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.12
murmurhash==1.0.10
mypy-extensions==0.4.3
mysqlclient==2.2.4
narwhals==1.21.1
netifaces==0.11.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nltk==3.8.1
numpy==2.2.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
odfpy==1.4.1
olefile==0.46
openai==1.107.3
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
parse==1.20.2
pathlib==1.0.1
pathspec==0.9.0
pdf2image==1.17.0
pdfminer.six==20231228
peewee==3.18.2
pexpect==4.8.0
pillow==11.1.0
platformdirs==4.3.6
playwright==1.53.0
preshed==3.0.9
prettytable==3.16.0
propcache==0.2.1
protobuf==5.29.2
prov==2.0.1
psutil==5.9.0
psycopg2==2.9.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
puremagic==1.28
pyarrow==18.1.0
pycairo==1.20.1
pycparser==2.22
pycryptodome==3.20.0
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
pyee==13.0.0
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.25.3
PyMySQL==1.1.2
PyNaCl==1.5.0
pypandoc==1.15
pyparsing==3.2.1
PyPDF2==3.0.1
pyppeteer==2.0.0
pyquery==2.0.1
pyRFC3339==1.1
pyrsistent==0.20.0
PySocks==1.7.1
pytesseract==0.3.13
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==3.3.0
python-slugify==8.0.4
pytz==2024.2
pyxdg==0.27
pyxnat==1.6.3
PyYAML==5.4.1
pyzbar==0.1.9
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
requests-html==0.10.0
rest-framework-simplejwt==0.0.2
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.3
scipy==1.15.2
SecretStorage==3.3.1
selenium==4.30.0
simplejson==3.20.1
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.1
srsly==2.4.8
ssh-import-id==5.11
starlette==0.46.1
streamlit==1.41.1
stripe==2.60.0
sympy==1.14.0
systemd-python==234
tablib==3.5.0
tenacity==9.0.0
text-unidecode==1.3
thinc==8.2.3
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==1.2.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
transformers==4.52.3
trio==0.28.0
trio-websocket==0.11.1
triton==3.3.0
typer==0.9.4
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
UNKNOWN @ git+ssh://git@github.com/VyLabs-AI/resume_lookbook.git@053d21d057e456afcc574811b94b98b40e057713
uritemplate==4.1.1
urllib3==1.26.20
usb-creator==0.3.7
uvicorn==0.34.0
virtualenv==20.27.1
w3lib==2.3.1
wadllib==1.3.6
wasabi==1.1.2
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.3.4
websocket-client==1.8.0
websockets==15.0.1
whitenoise==5.3.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.18.3
yfinance==0.2.66
zipp==1.0.0
